Allow to specifies multiple protocols to be tested (#5)
## What does this PR do ?

This PR enables us to specifies multiples protocols that the snippets should be tested with.
The snippet testing framework will then run the snippet for each different protocols that should be tested.

In the config.yml the snippets part has a new `protocols` field that can be used to defined the defaults protocols that the snippets should be tested with.
If not present this is defaulted to `websocket` only.

**Example of config.yml**
```yml
snippets:
  mount: /mnt
  path: doc/**/snippets/*.test.yml
  templates: /mnt/test/templates
  dest: /var/snippets
  protocols:
    - websocket
    - http
```

Each snippet can override the default configuration of the `config.yml` file by specifying the protocols themselves
**Example check-rights.test.yml**
```yml
name: auth#checkRights
description: <...>
hooks:
  before: <...>
  after: <...>
template: print-result
expected: true
protocols:
  - http
```
This tests will only be run for `http` while all the other that doesn't specifies their procotols will be run with the default `websocket` and `http`  from the previous config.

This PR also adds a new variable `{{ snippet.protocol }}` that will be replaced by the protocol that should be used to run the snippet.
This variable can be used to define an environment variable in the command that should run the snippet in the `config.yml`
**Example of config.yml**
```yml
    dart:
      service: <...>
      path: <...>
      build:
        cmd: <...>
      run:
        before: <...>
        cmd: SNIPPET_PROTOCOL={{ snippet.protocol }} dart {{ snippet.source }}
```
Like so for each protocol we will have a `SNIPPET_PROTOCOL` environment variable that is equal to the protocol that should be used. This environment variable can then be used in the templates to determine which protocol should be used.